Move local location backup setting (#23548)
