+++ src/panels/config/backup/ha-config-backup-settings.ts
@@ -1,15 +1,21 @@
+import { mdiDotsVertical, mdiHarddisk } from "@mdi/js";
 import type { PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
 import { isComponentLoaded } from "../../../common/config/is_component_loaded";
 import { fireEvent } from "../../../common/dom/fire_event";
+import { shouldHandleRequestSelectedEvent } from "../../../common/mwc/handle-request-selected-event";
 import { debounce } from "../../../common/util/debounce";
 import { nextRender } from "../../../common/util/render-status";
 import "../../../components/ha-button";
+import "../../../components/ha-button-menu";
 import "../../../components/ha-card";
+import "../../../components/ha-icon-button";
 import "../../../components/ha-icon-next";
+import "../../../components/ha-list-item";
 import "../../../components/ha-alert";
 import "../../../components/ha-password-field";
+import "../../../components/ha-svg-icon";
 import type { BackupConfig } from "../../../data/backup";
 import { updateBackupConfig } from "../../../data/backup";
 import type { CloudStatus } from "../../../data/cloud";
@@ -21,6 +27,7 @@ import type { BackupConfigData } from "./components/config/ha-backup-config-data
 import "./components/config/ha-backup-config-encryption-key";
 import "./components/config/ha-backup-config-schedule";
 import type { BackupConfigSchedule } from "./components/config/ha-backup-config-schedule";
+import { showLocalBackupLocationDialog } from "./dialogs/show-dialog-local-backup-location";
 
 @customElement("ha-config-backup-settings")
 class HaConfigBackupSettings extends LitElement {
@@ -94,6 +101,28 @@ class HaConfigBackupSettings extends LitElement {
         .narrow=${this.narrow}
         .header=${"Backup settings"}
       >
+        ${isComponentLoaded(this.hass, "hassio")
+          ? html`
+              <ha-button-menu slot="toolbar-icon">
+                <ha-icon-button
+                  slot="trigger"
+                  .label=${this.hass.localize("ui.common.menu")}
+                  .path=${mdiDotsVertical}
+                ></ha-icon-button>
+                <ha-list-item
+                  graphic="icon"
+                  @request-selected=${this._changeLocalLocation}
+                >
+                  <ha-svg-icon
+                    slot="graphic"
+                    .path=${mdiHarddisk}
+                  ></ha-svg-icon>
+                  Change default action location
+                </ha-list-item>
+              </ha-button-menu>
+            `
+          : nothing}
+
         <div class="content">
           <ha-card id="schedule">
             <div class="card-header">Automatic backups</div>
@@ -164,6 +193,14 @@ class HaConfigBackupSettings extends LitElement {
         </div>
       </hass-subpage>
     `;
+  }
+
+  private async _changeLocalLocation(ev) {
+    if (!shouldHandleRequestSelectedEvent(ev)) {
+      return;
+    }
+
+    showLocalBackupLocationDialog(this, {});
   }
 
   private _scheduleConfigChanged(ev) {
